Extract sync lifecycle actions
